// packages/archive/src/proto/Chunk.ts
export interface PBChunk {
  serializationFormatVersion: string;
  /** Interned strings */
  internedStrings: string[];
  internedMetaPointers: PBMetaPointer[];
  /** Interned languages */
  internedLanguages: PBLanguage[];
  nodes: PBNode[];
}

export interface PBLanguage {
  siKey?: number | undefined;
  siVersion?: number | undefined;
}

export interface PBNode {
  /** Yes, we should _not_ have null IDs, but if necessary we can represent them */
  siId?: number | undefined;
  mpiClassifier: number;
  properties: PBProperty[];
  containments: PBContainment[];
  references: PBReference[];
  /** This is a list of indexes representing the string values corresponding to the IDs of the annotations */
  siAnnotations: number[];
  siParent?: number | undefined;
}

export interface PBMetaPointer {
  /**
   * Instead of duplicating language+version as raw numbers everywhere,
   * point to the languages table.
   */
  liLanguage: number;
  /** concept/key within that language */
  siKey?: number | undefined;
}

export interface PBProperty {
  mpiMetaPointer: number;
  siValue?: number | undefined;
}

export interface PBContainment {
  mpiMetaPointer: number;
  siChildren: number[];
}

export interface PBReference {
  mpiMetaPointer: number;
  values: PBReferenceValue[];
}

export interface PBReferenceValue {
  siResolveInfo?: number | undefined;
  siReferred?: number | undefined;
}

function createBasePBChunk(): PBChunk {
  return {
    serializationFormatVersion: '',
    internedStrings: [],
    internedMetaPointers: [],
    internedLanguages: [],
    nodes: [],
  };
}

export const PBChunk = {
  encode(message: PBChunk, writer: _m0.Writer = _m0.Writer.create()): _m0.Writer {
    if (message.serializationFormatVersion !== '') {
      writer.uint32(10).string(message.serializationFormatVersion);
    }
    for (const v of message.internedStrings) {
      writer.uint32(18).string(v!);
    }
    for (const v of message.internedMetaPointers) {
      PBMetaPointer.encode(v!, writer.uint32(26).fork()).ldelim();
    }
    for (const v of message.internedLanguages) {
      PBLanguage.encode(v!, writer.uint32(34).fork()).ldelim();
    }
    for (const v of message.nodes) {
      PBNode.encode(v!, writer.uint32(42).fork()).ldelim();
    }
    return writer;
  },

  decode(input: _m0.Reader | Uint8Array, length?: number): PBChunk {
    const reader = input instanceof _m0.Reader ? input : _m0.Reader.create(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBasePBChunk();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1:
          if (tag !== 10) {
            break;
          }

          message.serializationFormatVersion = reader.string();
          continue;
        case 2:
          if (tag !== 18) {
            break;
          }

          message.internedStrings.push(reader.string());
          continue;
        case 3:
          if (tag !== 26) {
            break;
          }

          message.internedMetaPointers.push(PBMetaPointer.decode(reader, reader.uint32()));
          continue;
        case 4:
          if (tag !== 34) {
            break;
          }

          message.internedLanguages.push(PBLanguage.decode(reader, reader.uint32()));
          continue;
        case 5:
          if (tag !== 42) {
            break;
          }

          message.nodes.push(PBNode.decode(reader, reader.uint32()));
          continue;
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skipType(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): PBChunk {
    return {
      serializationFormatVersion: isSet(object.serializationFormatVersion)
        ? globalThis.String(object.serializationFormatVersion)
        : '',
      internedStrings: globalThis.Array.isArray(object?.internedStrings)
        ? object.internedStrings.map((e: any) => globalThis.String(e))
        : [],
      internedMetaPointers: globalThis.Array.isArray(object?.internedMetaPointers)
        ? object.internedMetaPointers.map((e: any) => PBMetaPointer.fromJSON(e))
        : [],
      internedLanguages: globalThis.Array.isArray(object?.internedLanguages)
        ? object.internedLanguages.map((e: any) => PBLanguage.fromJSON(e))
        : [],
      nodes: globalThis.Array.isArray(object?.nodes)
        ? object.nodes.map((e: any) => PBNode.fromJSON(e))
        : [],
    };
  },

  toJSON(message: PBChunk): unknown {
    const obj: any = {};
    if (message.serializationFormatVersion !== '') {
      obj.serializationFormatVersion = message.serializationFormatVersion;
    }
    if (message.internedStrings?.length) {
      obj.internedStrings = message.internedStrings;
    }
    if (message.internedMetaPointers?.length) {
      obj.internedMetaPointers = message.internedMetaPointers.map((e) => PBMetaPointer.toJSON(e));
    }
    if (message.internedLanguages?.length) {
      obj.internedLanguages = message.internedLanguages.map((e) => PBLanguage.toJSON(e));
    }
    if (message.nodes?.length) {
      obj.nodes = message.nodes.map((e) => PBNode.toJSON(e));
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<PBChunk>, I>>(base?: I): PBChunk {
    return PBChunk.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<PBChunk>, I>>(object: I): PBChunk {
    const message = createBasePBChunk();
    message.serializationFormatVersion = object.serializationFormatVersion ?? '';
    message.internedStrings = object.internedStrings?.map((e) => e) || [];
    message.internedMetaPointers =
      object.internedMetaPointers?.map((e) => PBMetaPointer.fromPartial(e)) || [];
    message.internedLanguages =
      object.internedLanguages?.map((e) => PBLanguage.fromPartial(e)) || [];
    message.nodes = object.nodes?.map((e) => PBNode.fromPartial(e)) || [];
    return message;
  },
};

function createBasePBLanguage(): PBLanguage {
  return { siKey: undefined, siVersion: undefined };
}

export const PBLanguage = {
  encode(message: PBLanguage, writer: _m0.Writer = _m0.Writer.create()): _m0.Writer {
    if (message.siKey !== undefined) {
      writer.uint32(8).uint32(message.siKey);
    }
    if (message.siVersion !== undefined) {
      writer.uint32(16).uint32(message.siVersion);
    }
    return writer;
  },

  decode(input: _m0.Reader | Uint8Array, length?: number): PBLanguage {
    const reader = input instanceof _m0.Reader ? input : _m0.Reader.create(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBasePBLanguage();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1:
          if (tag !== 8) {
            break;
          }

          message.siKey = reader.uint32();
          continue;
        case 2:
          if (tag !== 16) {
            break;
          }

          message.siVersion = reader.uint32();
          continue;
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skipType(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): PBLanguage {
    return {
      siKey: isSet(object.siKey) ? globalThis.Number(object.siKey) : undefined,
      siVersion: isSet(object.siVersion) ? globalThis.Number(object.siVersion) : undefined,
    };
  },

  toJSON(message: PBLanguage): unknown {
    const obj: any = {};
    if (message.siKey !== undefined) {
      obj.siKey = Math.round(message.siKey);
    }
    if (message.siVersion !== undefined) {
      obj.siVersion = Math.round(message.siVersion);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<PBLanguage>, I>>(base?: I): PBLanguage {
    return PBLanguage.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<PBLanguage>, I>>(object: I): PBLanguage {
    const message = createBasePBLanguage();
    message.siKey = object.siKey ?? undefined;
    message.siVersion = object.siVersion ?? undefined;
    return message;
  },
};
// ...
export const PBMetaPointer = {
  encode(message: PBMetaPointer, writer: _m0.Writer = _m0.Writer.create()): _m0.Writer {
    if (message.liLanguage !== 0) {
      writer.uint32(8).uint32(message.liLanguage);
    }
    if (message.siKey !== undefined) {
      writer.uint32(16).uint32(message.siKey);
    }
    return writer;
  },

  decode(input: _m0.Reader | Uint8Array, length?: number): PBMetaPointer {
    const reader = input instanceof _m0.Reader ? input : _m0.Reader.create(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBasePBMetaPointer();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1:
          if (tag !== 8) {
            break;
          }

          message.liLanguage = reader.uint32();
          continue;
        case 2:
          if (tag !== 16) {
            break;
          }

          message.siKey = reader.uint32();
          continue;
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skipType(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): PBMetaPointer {
    return {
      liLanguage: isSet(object.liLanguage) ? globalThis.Number(object.liLanguage) : 0,
      siKey: isSet(object.siKey) ? globalThis.Number(object.siKey) : undefined,
    };
  },

  toJSON(message: PBMetaPointer): unknown {
    const obj: any = {};
    if (message.liLanguage !== 0) {
      obj.liLanguage = Math.round(message.liLanguage);
    }
    if (message.siKey !== undefined) {
      obj.siKey = Math.round(message.siKey);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<PBMetaPointer>, I>>(base?: I): PBMetaPointer {
    return PBMetaPointer.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<PBMetaPointer>, I>>(object: I): PBMetaPointer {
    const message = createBasePBMetaPointer();
    message.liLanguage = object.liLanguage ?? 0;
    message.siKey = object.siKey ?? undefined;
    return message;
  },
};
// ...
type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin
  ? T
  : T extends globalThis.Array<infer U>
    ? globalThis.Array<DeepPartial<U>>
    : T extends ReadonlyArray<infer U>
      ? ReadonlyArray<DeepPartial<U>>
      : T extends {}
        ? { [K in keyof T]?: DeepPartial<T[K]> }
        : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin
  ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };